Move Gateway StarkNet Adapter to its own package
